Applications::AudioDecay: Use AudioProcessorEditor for UI
