+++ code/dsp/processors/AudioDecayProcessor.h
@@ -1,19 +1,17 @@
 #pragma once
 
 #include "JuceHeader.h"
+#include "../../ui/CustomLookAndFeel.h"
 
 //==============================================================================
 class AudioDecayProcessor  : public juce::AudioProcessor
 {
 public:
+    // TODO: Make this private
+    AudioProcessorValueTreeState state;
+    
     //==============================================================================
     AudioDecayProcessor();
-    
-    //==============================================================================
-    
-    void setQuantisationLevel(int bitDepth);
-    void setDownsampleFactor(int downsampleFactor);
-    void setWetDryMix(float mix);
 
     //==============================================================================
     bool isBusesLayoutSupported (const BusesLayout& layouts) const override;
@@ -24,7 +22,7 @@ public:
     void processBlock (juce::AudioBuffer<float>& buffer, juce::MidiBuffer&) override;
     
     //==============================================================================
-    juce::AudioProcessorEditor* createEditor() override            { return new juce::GenericAudioProcessorEditor (*this); }
+    juce::AudioProcessorEditor* createEditor() override            { return new AudioDecayPluginProcessorEditor (*this); }
     bool hasEditor() const override                          { return true; }
     const String getName() const override                    { return "AudioDecay"; }
     bool acceptsMidi() const override                        { return false; }
@@ -42,6 +40,68 @@ public:
     void setStateInformation (const void* data, int sizeInBytes) override;
 
 private:
+    //==============================================================================
+    class AudioDecayPluginProcessorEditor
+    : public juce::AudioProcessorEditor
+    {
+    public:
+        
+        AudioDecayPluginProcessorEditor(AudioDecayProcessor& owner)
+        : juce::AudioProcessorEditor(owner)
+        , mBitDepthSlider("Bit-Depth", "bit")
+        , mDownsamplingSlider("Downsampling", "x")
+        , mWetDrySlider("Mix", "")
+        , mBitDepthAttachment(owner.state, "quantisation", mBitDepthSlider)
+        , mDownsamplingAttachment(owner.state, "downsampling", mDownsamplingSlider)
+        , mWetDryAttachment(owner.state, "wetdry", mWetDrySlider)
+        {
+            addAndMakeVisible(&mBitDepthSlider);
+            mBitDepthSlider.mLabels.add({0.0f, "3"});
+            mBitDepthSlider.mLabels.add({1.0f, "24"});
+            
+            addAndMakeVisible(&mDownsamplingSlider);
+            mDownsamplingSlider.mLabels.add({0.0f, "1"});
+            mDownsamplingSlider.mLabels.add({1.0f, "10"});
+            
+            addAndMakeVisible(&mWetDrySlider);
+            mWetDrySlider.mLabels.add({0.0f, "Dry"});
+            mWetDrySlider.mLabels.add({1.0f, "Wet"});
+            setSize (600, 200);
+        }
+        
+        ~AudioDecayPluginProcessorEditor() override {}
+        
+        void paint(juce::Graphics& g) override
+        {
+            g.fillAll (getLookAndFeel().findColour (juce::ResizableWindow::backgroundColourId));
+        }
+        
+        void resized() override
+        {
+            auto bounds = getLocalBounds();
+            bounds.reduced(20, 20);
+            
+            auto constexpr numUIElements = 3;
+            auto const rotaryWidth = static_cast<int>(bounds.getWidth() * 1 / numUIElements);
+            auto const spacing = (bounds.getWidth() - rotaryWidth * numUIElements) / (numUIElements - 1);
+            
+            mBitDepthSlider.setBounds(bounds.removeFromLeft(rotaryWidth));
+            bounds.removeFromLeft(spacing);
+            mDownsamplingSlider.setBounds(bounds.removeFromLeft(rotaryWidth));
+            bounds.removeFromLeft(spacing);
+            mWetDrySlider.setBounds(bounds.removeFromLeft(rotaryWidth));
+        }
+        
+    private:
+        RotarySliderWithLabels mBitDepthSlider;
+        RotarySliderWithLabels mDownsamplingSlider;
+        RotarySliderWithLabels mWetDrySlider;
+        
+        juce::AudioProcessorValueTreeState::SliderAttachment mBitDepthAttachment;
+        juce::AudioProcessorValueTreeState::SliderAttachment mDownsamplingAttachment;
+        juce::AudioProcessorValueTreeState::SliderAttachment mWetDryAttachment;
+    };
+    
     //==============================================================================
     int mBlockSize;
     int mSampleRate;
